V 1.5.3
The latest version of the BDB-Compatible Gemini Paraglider has remedied the major problems that plagued the initial release version, whose download links had to be closed due to a major problem with the part. With the help of some of the most esteemed modders of the KSP community, I, along with a host of other modders, including but not limited to Pak, steed_crudgeon, theRagingIrishman, and Angel-125, have finally gotten the paraglider wing to properly work. Enjoy the latest, ready-for-release version!