V 1.0 "It Doesn't Control Like a Pig Anymore!"
Completely re-jiggered the paraglider so it has more lift and is also now a control surface as well as a wing, so it automatically counters what used to be automatic rolling. IMPORTANT NOTE: Be very gentle when banking; Simply tapping "Q" or "E" will automatically make the wing slowly turn your craft around, then re-orient it to fly straight again.